CP-5596: Have VM_start fail if vgpu fails to start
Factored out the code to start Qemu into a function that forks a daemon
and waits for it to finish initialising (with a write to XenStore). This
has then be reused to kick off the vgpu emulator binary which can now
fail in the same way if it doesn't initialise. vgpu is kicked off in
conjuction with Qemu since it's part of the device model.

Imported from xen-api.git 620069e60130e35eba8823c7538eb72fb3b36940

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>
Imported-by: John Else <john.else@citrix.com>