CP-8378: Use raw VDI for suspend images
This patch removes the Ext filesystem from the suspend VDI and instead uses
a raw VDI. This avoids the need to mount the device in dom0. Restoring from
older suspend image VDIs (with a filesystem present) remain to work since we
check for the presence of the the Ext magic value and reverts to mounting as
before if present.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>