Support migration from pre-levelling-v2 hosts
When importing VM metadata that was sent to us by a xenopsd running on another
host (e.g. during a live migration), we need to ensure that the VM has a valid
featureset configured in its platformdata.

If platformdata does not contain a featureset, then we are importing a VM that
comes from a levelling-v1 host. In this case, give it a featureset that
contains all features that this host has to offer.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>