CA-191040: Ensure vm's frontend's backend key is correct
A malicious guest can get the vdi id/uuid, qos and vbd device of other guests
by replacing its frontend's backend key with the backend key of other guests.
This patch tries to put a defensive check to make sure that the backend key
maps to the vm that is trying to read it.

Signed-off-by: Koushik Chakravarty <koushik.chakravarty@citrix.com>