CA-220466: bootloader: check kernel/ramdisk paths
When file-paths are returned from the bootloader (pygrub or eliloader)
to be used for the guest's kernel or ramdisk (initrd), now we check
that the file-paths point to files in the expected directories.

This is to add some protection in case a malicious guest were to trick
pygrub or eliloader into specifying a guest-chosen file elsewhere in
the dom0 filesystem.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>
(cherry picked from commit f01da7c268e289fde14ae55c4c14920470f35eb4)