CA-225067: move qemu-dm to default cgroup slice
This avoids the need to rely on the cgrulesengd daemon to move the process into
its designated slice. This daemon fails to do its job if it takes too long to
write to the tasks file. On a 4.4 kernel, it can take several milliseconds to
write to this file, so that kind of failure is very likely.

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>