CA-227605: issues with PVS caching under stress tests
Ticket title: VM not cached after start/resume in PVS Stress tests -
missing OVS rules for intercepting traffic.

In the previous update (CA-227626) we used `xenstore-ls` to get the
appropriate path for the proxy state. This fails if a key is removed
from xenstore while reading. This patch swaps `xenstore-ls` with the
safer `xenstore-list` at the price of a slightly more complicated code.

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>