CP-20761: PPX
Switch to newer libraries using ppx, and use cppo instead of optcomp,
because optcomp is a camlp4 syntax extension.

And also update files generated by OASIS by running "oasis setup"

This is a manual rebase of the ppx-ely branch on top of master, and
updates more files that may not be compiled by default, as I used sed to
update the old camlp4 syntax to the new ppx one.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>