PPX
Manually rebase the ppx-ely branch on top of master. I used sed to
update to the Xpath library names, and the new ppx syntax, so this
updates more things that the original a2697af commit: it also updates
the files in the libvirt and qemu directories, but these are probably
not built by default.

And also update files generated by OASIS by running "oasis setup"

This commit matches the automatic git rebase, except for the two
directories mentioned above, that were additionally updated.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>