CP-20603: Pass vGPU fd to xenguest on source host during migration
When starting a migration of a VM with a vGPU, we need to tell xenguest about
the file descriptor for the connection that was opened for transferring the
vGPU state to the destination host. Xenguest will pass this on to DEMU, which
kicks off the vGPU migration phase.

Do this by adding a new optional parameter to the `VM_save` operation for the
additional file descriptor.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>