Do entire vGPU state migration inside VM_{save,restore}_vgpu operations
...and do this in parallel with the `VM_{save,restore}` operations.

Previously, `VM_{save,restore}_vgpu` would only kick off the migration process,
and we would rely on something else to sync up when the process completes. This
turned out to be the wrong approach. Furthermore, it is not necessary to start
DEMU on the destination before starting the process on the source, so we don't
need to have separate operations for the setup and actual migration phases.

We can conveniently use the new `Parallel` operation for this purpose.

Finally, DEMU now writes its status to xenstore, which we use to keep track of
its progress and any errors.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>