CP-17693: take action when qemu crashes or exits unexpectedly
The HVM guest is very likely to hang or crash when qemu crashes, but the VM
will usually stay running until a force shutdown is applied. This patch uses
the forkexecd waitpid facility in an asynchronous way to signal to xenopsd
that qemu crashed or ended unexpectedly after the VM started. A signal masking
system disables the signal if an expected domain destroy event is detected, to
prevent spurious asynchronous signals propagating during or after a normal
shutdown.

Depending on the value of a new xenopsd.conf option action-after-qemu-crash,
xenopsd will take one of the following actions after receiving this signal:
- poweroff: xenopsd will power off the VM
- pause: xenopsd will pause the VM
- otherwise, xenopsd reverts to its previous behaviour of not waiting for any
qemu signal and not taking any action automatically.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>