CA-254911: delay shutdown cleanup actions to mitigate a potential stack overflow
This commit changes the behaviour of the cleanup actions using `deferred_operation` instead
of `immediate_operation`. This pushes the operation on a thread queue instead of performing
them immediately, mitigating the issue observed in CA-254911. Although the stack overflow
has not been seen in the tests, I think with this PR we add a mitigation but I do not think
we can guarantee that the stack overflow is always fixed: this can depend on how fast the
queue of operations builds up.

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>