Revert "Hack: hotplug-state does somehow not get removed for qdisk"
Because I've now made qemu remove hotplug-state when the blkif ring is closing.
If we don't revert this hack then the xenstore trees get torn down without
allowing the frontend/backend states to advance to 6/Closing, so blk_disconnect
is not called in qemu, so the SM layer's call x-blockdev-del in deactivate()
doesn't succeed.

This reverts commit 3a665caefbb1679b4dc3ef16d7479cfaf4546374.