CP-22500: support clean suspend, shutdown etc for HVM Linux guests
qemu-trad would detect that a HVM Linux guest PV driver was running
by checking if the PV driver product_id was 3:

qemu-trad.pg:hw/xen_platform.c:
++
++      if (product_id == 3 && build_number <= 0xff &&
++                !drivers_blacklisted)
++                xenstore_notify_unplug();
+     }

were 3 is the product_id as defined in Linux' xen/platform_pci.h:
https://github.com/torvalds/linux/blob/5924bbecd0267d87c24110cbe2041b5075173a25/include/xen/platform_pci.h#L5

  #define XEN_IOPORT_LINUX_PRODNUM 0x0003

Forward-porting this detection method to upstream QEMU was NACKed.

So, after some discussion in CP-22500, we decided to base this decision
on the VM template.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>