Send restore:xenguest in restore_libxc_record
The new protocol with emu-manager is for xenopsd to request restoring the next
bit from the file using the new "restore:" command followed by the name of the
thing to be restored, which is determined by the last read header in the
stream. First up is "restore:xenguest" for the "libxc" record.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>