CP-23595: Introduce device model profiles to xenopsd - Stage 1
For Stage 1:
* Always use qemu-trad unless a VM is specifically set to
use one of the known profiles:
  - qemu-trad
  - qemu-upstream
  - qemu-upstream-compat
* The device model profile used if VM.platform["device-model"] is
not set shall be 'qemu-trad' (the fallback)
* Each profile has its own device model wrapper.

The device-model key in VM.platform was white-listed by Dave in xen-api
commit 63d41a86fbbda3a384f41f97567222d4e0afd013. The initial implementation
in xenopsd used the "Xenops_server_xen.choose_alternative()" method of
using the filesystem to choose the binary associated with a white-listed
value. We do not use this alternative() function to select between a device-model
value and its binary because the qemu wrappers and binaries are in arbitrary
locations in dom0. Instead, we use a new Profile module in Device.Dm for managing
different qemu profiles (qemu-trad, qemu-upstream-compat, qemu-upstream). This
Profile module will be extended in the future to handle more differences between
types of device models.

In the tests, Resources was able to find the executables for both qemu-dm-wrapper
(qemu-trad) and upstream-compat-qemu-dm-wrapper (upstream qemu for Stage 1):

[debug||0 ||xenopsd-xc] qemu-dm-wrapper = /usr/libexec/xenopsd/qemu-dm-wrapper (path to the qemu-dm-wrapper script)
[debug||0 ||xenopsd-xc] upstream-compat-qemu-dm-wrapper = /usr/lib64/xen/bin/qemu-wrapper (path to the upstream compat qemu-dm-wrapper script)

and the VM started properly as expected.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>