Turn --restore option of DEMU into a flag; drop --resumefd
The restore fd does not need to be specified on the command line anymore,
because emu-manager passes on the fd instead. All xenopsd needs to do is start
DEMU in restore mode.

There is an existing --restore option that used to take a file name, which is
no longer used. We repurpose this to be the restore flag.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>