CP-23986: Implement QMP event thread in xenopsd
Create a qmp_event_thread to monitor QMP events of all QEMU-upstream
domains and handle necessary events.

Also create a pipe as an extra fd, when a QEMU-upstream based HVM is
started, write a character to the pipe to wake up qmp_event_thread.

We use Core.Linux_ext.Epoll in qmp_event_thread because it is already
part of xs-opam.

Module Lookup is used to abstract two hashtables (fd to domainid and
domain to qmp connection) in a single data structure that's easy to use.

Module Monitor is used to abstract epoll operations and filter input
events.

Signed-off-by: Liang Dai <liang.dai1@citrix.com>