Wait for DEMU to reach initialising or running
Previously, we'd wait until DEMU had reached the "running" state, which is fine
for migrations, when DEMU was started on a separate thread. For migrations, we
now start DEMU in VM_receive_memory on the main receiver thread, where the
following operation (VM_restore) actually causes DEMU to restore and reach the
"running" state. Therefore, starting DEMU should not block. The same applies to
VM_resume.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>