CA-264331: Make shutdown ack timeout configurable
This setting was previously configurable from the `xapi.conf` file,
when xenopsd was still a part of xapi. When they split, the setting
was not transferred and nor was it wired up to be used by xenopsd.

This commit moves that setting into the `xenopsd.conf` file and uses
it in the domain shutdown request when it is provided, otherwise it
uses the existing default value of 60 seconds.

Signed-off-by: Frederico Mazzone <frederico.mazzone@citrix.com>