CP-24361: use the qemu backends for init_daemon
The functions start_vgpu and __start were not using the available qemu
backends to run init_daemon. This patch moves these 2 functions (and the
corresponding dependencies start, restore and start_vnc_only) after the
dispatched version of init_daemon, so that QMP_Event.add can be called
as expected when the qemu-upstream backend is used.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>