CA-269046: add "console/limit" key to xenstore for a domain
When we don't have this key present xenconsoled tends to
accumulate the output of guest PV consoles indefinitely if there
are no clients to consume it in Dom0. The problem manifests itself
when "disable_pv_vnc" parameter in other-config is set to 1 which
means there are no vncterm clients started for PV consoles.

Set this key to a reasonable limit of 64KB that will tell
xenconsoled to rotate the internal buffer instead of continously
increasing it.

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>