CA-265581: pass pipe_r to emu_manager, as is done in xenguest helper
When upgrading from XS6.2, a conversion function needs to be used to
convert the VM save stream to the new format.

- Originally, the conversion function would send a pipe fd to xenguest.
- After the introduction of the emu-manager code, a copy-paste error
prevented the pipe from being passed to emu-manager (which will pass
it along to xenguest), causing a VM migration from XS 6.2 to fail.

This patch fixes the issue by reusing the proper pipe fd to connect the
conversion function to emu-manager. It does so by replacing main_fd in
the call to with_emu_manager for pipe_r. It's not necessary to modify
vgpu_fd in this case because XS 6.2 will never have vgpu_fds.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>