CA-270640: Avoid leaking waker threads
If we receive unexpected responses from emu-manager, the
handle_result loop exits immediately, potentially leaving
"requests" threads blocked and waiting to be woken up.

This can be avoided by waking up all the remaining threads
in a finalizer loop.

This does not solve the potential leak of the reader threads.

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>