CA-282008 make qmp interaction more robust
qmp_write() does not check the result of a qmp command and also does not
pass it on to the caller such that errors get lost. This commit
implements a check of the result of qmp_write. In case of a failure, a
new exception QMP_Error is raised.

qmp_write also detects return messages that carry data that is ignored.
This indicates that qmp_write probably was not used appropriately and
qmp_write_and_read should have been used. This is logged but not flagged
as an error.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>