CA-282008 introduce qmp_cmd, qmp_safe_cmd
Functions qmp_cmd and qmp_safe_cmd send a command to QEMU, wait for the
response, and return it. The safe variant checks that the result is
Success and otherwise raises an exception.

Both functions discard messages received from QEMU that are not
identified as the result for the command that has been sent.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>