CP-27216: uid/gid save and restore of device file
This ticket is to make the uid/gid more accurate, as uid/gid of almost all the
usb devices in the path (eg. "/dev/bus/usb/001/003") is 0/0. But it
still can be other values. So we will save the uid/gid of the device file
when plugging it to vm and restore it in unplug step.

Signed-off-by: Taoyong Ding <taoyong.ding@citrix.com>