CP-27581: Move MMIO hole size key to a persistent location
This complements the changes to xenguest and hvmloader which in complex
serves as additional protection from unpredictable MMIO hole size
fluctuations that might occur for a variety of reasons. Although,
it's expected that all migration streams starting from Kolkata
will have this key, we still need to guess the size of MMIO hole for
the VMs from the previous releases (Jura and older) and handle them as
default cases - hence there is still the special handling for legacy
and vGPU VMs.

Signed-off-by: Igor Druzhinin <igor.druzhinin@citrix.com>