Revert "CP-27560 set device model fallback to qemu-upstream-compat"
This reverts commit c5df2a00a992a40e1a1f4af6c8b690f708a36aa9.

The problem is the use of Device.Profile.fallback in xenopsd: suspend()
calls suspend_emu_manager() which selects the backend based on the
device model profile. A PV guest does not have an assigned device model
profile and uses Device.Profile.fallback, which has been QEMU trad so
far. But switching this to QEMU compat cause xenopsd looking for QMP -
which is wrong.  The overall problem is that the QEMU trad backend also
handles PV guests but this needs to become independent from it to
eliminate QEMU trad.