Fix for qemu datapath which uses NBD
We still use the params key for the NBD uri, because that's what the
qemu-wrapper script and probably other places in xenopsd expect.
Now SMAPIv3 should tell us the complete NBD uri, including the
exportname, therefore we don't need to add it in qemu-wrapper.
We try to insert the qemu-params key when required for qemuback.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>