CA-290644: introduce a version number in the xenopsd persistent_t vm metadata
Xapi is responsible for many pool VM lifecycle operations such as vm-start,
shutdown,import,export,snapshot etc and it uses the Xapi DB to store information
that will be used during these operations. To support moving this information
across different Xapi versions (eg. during a host update), a versioning system
in the Xapi DB is used to take specific db_upgrade actions for each VM depending
on both the old and the new Xapi DB versions involved in the update.

Xenopsd is responsible for all the host VM lifecycle operations that occur while
the VM is in the running state, such as vm-migrate,suspend,resume etc and it
uses the Xenopsd DB to persist the running VM metadata (even across hosts) until
the VM shuts down. However, unlikely the Xapi DB, the Xenopsd persistent DB does
not have a versioning field, so it can't take specific actions depending on the
version of Xenopsd that the VM was started on and the current version of Xenopsd
that the VM is attempting to migrate or resume to.

Here we introduce a new version field to the Xenopsd DB persistent VM metadata,
which will be useful to support specific Xenopsd actions during cross-pool
migration and cross-version suspend/resume operations:

- The default value for this version field is 0, which is going to be used to
represent that a VM was started using any Xenopsd version without this field.

- The initial value for this version field is 1, which is going to be used to
indicate that the Xenopsd persistent data was used to start a VM using a Xenopsd
version that has the knowledge of a version field.

- In the future, the value for this field can be incremented further to
distinguish the Xenopsd persistent data from others created by earlier versions
of Xenopsd.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>