CA-290644: vm import_metadata: upgrade profile qemu-trad->qemu-upstream-compat
An existing xenopsd VM metadata record of a running VM is usually transferred
from another host using the import_metadata function on the destination host.
This function call set_internal_state to persist this VM metadata record on
the destination host.

This patch intercepts the received persistent record and applies any revisions
necessary to adapt the VM metadata record from an older host to the new host
as soon as this record is received and before it is persisted and used by any
other xenopsd atom or function of the destination host, including vm-migrate
and vm-resume.

The first revision of this kind, added by this patch, is the upgrade of the
qemu-trad profile metadata when the VM that has been started using a previous
version of xenopsd that has no persistent metadata version (which defaults to
version 0). This will not apply to VMs with qemu-trad profile migrating in a
pool where they were started with a persistent metadata version 1 or later.
This is by design, so that if a user manually reverts the VM profile in back
to qemu-trad in the xapi DB and restarts the VM, it will not be automatically
upgraded to the qemu-upstream-compat profile.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>