CP-24770: create QEMU disk command line in xenopsd instead of qemu-wrapper
This patch rebases commit ffddbbfb414d7a83ecde14f07fa0809941af5c9c
> CP-24770: Xenopsd calls now the qemu-wrapper script passing in the right
> parameters for disks and CD drives.
> Signed-off-by: Konstantina Chremmou <konstantina.chremmou@citrix.com>

In particular, it simplifies the change to the VmExtra.qemu_vbds field after
the VmExtra.non_persistent_t field was removed and qemu_vbds is available
now only as part of the VmExtra.persistent_t field, which needs to be
backwards-compatible with previous versions of xenopsd due to RPU and
cross-pool migration: instead of making qemu_frontend optional, which would
cause incompatibility with previous versions of the VmExtra.persistent_t
field, it adds an extra option 'Empty' to the qemu_frontend type. The
unmarshalling of this new option should be backwards-compatible for VMs
started in previous versions of xenopsd.

Dm_Common.qemu_args.disks', which creates disk parameters to qemu, is:

- removed from the Dm_Common backend, where it was not active due to the use
of qemu_disk_cmdline condition when creating the info.disks parameter in
xenops_server_xen.ml;

- not moved to the qemu-trad because qemu-trad does not use any disk
command line parameter, instead relying on xenstore and its qemu-wrapper;

- added to the qemu-upstream backend in an updated way as the result of
mapping the corresponding qemu disk parameters function in the qemu-upstream
wrapper, but without relying on xenstored and guesses since xenopsd contains
all the required information.

After the original commit above, a change for vbd, nbd, qdisk backends and
read-only disks was added to the qemu-wrapper. This further change will be
added in a following commit.

This patch:

- when plugging the vbds, now maps all the vbds to VM.persistent.qemu_vbds,
even the vbds that are an empty cdrom (which were never added to qemu_vbds
before this patch). The result is that Device.Dm.start now has the index
for the empty cdroms, needed to build the qemu-upstream command line.

- when unplugging the vbds, nows removes the vbd from VM.persistent.qemu_vbds,
without destroying the frontend of an empty cdrom. Tests in
https://github.com/xapi-project/xenopsd/pull/447/commits/ffddbbfb414d7a83ecde14f07fa0809941af5c9c
indicate that the behaviour calling create_frontend and destroy_frontend
is the same before and after this patch.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>