CA-291050: qemu: Don't unshare network namespace when hvm_serial is set
hvm_serial is often used for Windows kernel debugging and it typically
requires that QEMU uses the network (either actively making a connection
or passively waiting for a connection). If hvm_serial is set, don't
unshare the network namespace otherwise the common debugging use case is
broken.

This is not a security risk since preventing QEMU from using the network
is a defense-in-depth and hvm_serial should only be used for debugging.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>