CP-28663: build and fix suspend-image-viewer
The path to verify-stream-v2 was incorrect, preventing
suspend-image-viewer from showing the full suspend VDI's contents.

Also be explicit about the headers that are supported instead of using
wildcards in preparation for adding a new header type.

CP-29075: Make suspend-image-viewer parse /etc/xenopsd.conf to resolve the path to
verify-stream-v2.
This makes suspend-image-viewer work again, also added a small shell
script that shows how to use it.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>