CP-28663: implement VM.suspend/resume for UEFI
This adds a new Varstored record type to the suspend image.

Varstored gets stopped with a SIGTERM which causes it to write out all
the UEFI variables to the specified path.
We read back the file and put it into the suspend (migration) stream.
On resume we read the varstored record from the suspend image and write
it out to the resume path, which is picked up by varstored on resume.
We tell varstored to resume by explicitly passing --resume.

It is important to stop varstored at the correct place
(varstored exposes a PCI device to the guest):
* if we stop varstored after QEMU is killed then varstored won't be able
to reattach on resume because the device will be considered in use
* if we stop varstored while the guest is still running then upon resume
the PCI config space will be different (even if varstored attempts to restore it),
which means the guest is not able to access it anymore
* if we stop it when the guest is suspended but QEMU is still running
    then varstored is able to successfully reattach and a (Linux) guest
    able to successfully access the UEFI variables

We cannot immediately write out the varstored record when we stop it
(we would corrupt the migration stream by doing so), so we use the
efivars save path as a buffer to store the variables until we can.
This file (and all efi var related ones) will get deleted when we stop
qemu and the other daemons.

Stopping varstored will delete the xenstore pid key, so attempting to stop
it a second time (when qemu is stopped) is harmless.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>