CP-29054: use locked PID files for QEMU
Xenopsd currently relies on the PID stored in xenstore for killing qemu, vgpu and varstored.
But these daemons might've crashed on their own, which means that when we kill them we might kill some other random process that happens to have reused the PID.

The daemons should write their PID into a file and hold it locked (in addition to xenstore, which is useful for signaling when startup has completed successfully).

Qemu already supports this, so this patch adds a mode to DaemonMgmt that
uses pidfiles (if available) to determine which process to kill. To
support in-place upgrades missing pidfiles cause a fallback to reading
the pid from xenstore.

This doesn't completely remove the race condition (i.e. we can check the pidfile, see it is locked, meanwhile demu would die, and another process would reuse its PID, and we kill the wrong one, but it is very unlikely a PID wraparound would happen in such a short amount of time). The window for a race would be much narrower than if the daemon dies on VM start and we want to kill it possibly days later on VM stop.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>