CA-297602: Always create a physical-device node for HVM CD-ROMs
To increase performance, OVMF can use PV drivers. A previous change to
XAPI (9d0eb8f80507, CA-50383) prevented the VBD backend from being
created for HVM CD-ROMs due to some interaction with blkback.  However,
blktap3 is now used as the backend for CDs and there doesn't seem to be
any issue with it. Therefore, remove the code which prevents creating
the physical-device node to allow OVMF to use PV drivers.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>