CP-29622: UEFI overrides device-model to be qemu-upstream
Bios defaults to qemu-upstream-compat, but can use qemu-upstream if set
in device-model.

Our templates set qemu-upstream-compat even for UEFI, so unless we
change that we need this override.
The other option is to change the templates, but then we would need to
test and support 3 scenarios:
  * UEFI template, user clones and runs
    (device-model=qemu-upstream, firmare=uefi)
  * UEFI template, user clones, chooses BIOS mode from XenCenter, runs
    (device-model=qemu-upstream, firmware=bios)
  * BIOS template, user clones and runs
    (device-model=qemu-upstream-compat, firmware=bios)

With this override we only got 2 scenarios (the last 2 are equivalent):
  * UEFI template, user clones and runs
    (firmware=uefi, effectively device-model=qemu-upstream even though
     template says device-model=qemu-upstream-compat)
  * UEFI template, user clones chooses BIOS XenCenter and runs
    (firmware=bios, device-model=qemu-upstream-compat)
  * BIOS template, user clones and runs
    (firmware=bios, device-model=qemu-upstream-compat)

One could still manually set device-model=qemu-upstream, useful for
testing purposes but not necessarily something to support

Signed-off-by: Edwin Török <edvin.torok@citrix.com>