CP-24362: remove qemu-trad profile, treat it as qemu-upstream-compat
This removes the qemu-trad profile, and if we encounter
'platform:device-model=qemu-trad' or 'qemu-trad' from a migration/resume
we'll treat it as 'qemu-upstream-compat' instead.

We were already doing this for pre-Lima qemu-trad VMs, now we'll do it
for all qemu-trad VMs.

PV guests need to be handled differently because they do not have a Qemu associated with them.
Previously it would've defaulted to qemu-trad, which ignored errors from not being able to stop a non-existent qemu
(e.g. during a live migrate of a PV guest), but qemu-upstream-compat would fail here when it tries to send QMP commands.
Introduce a Qemu_none backend for PV guests that doesn't spawn a qemu.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>