CP-25680, CP-25605: do not attempt to start qemu upstream with qemu-trad arguments
When `platform:pvfb=true` is set Device.Dm.start_vnconly would attempt
to spawn a qemu, but if we deleted qemu-trad then it would attempt to
spawn qemu-upstream with the trad flags.
This won't work, and even if we adapt the flags to qemu-upstream the
`xenpv` machine model fails to start with:
`qemu-dm-28: Property '.max-ram-below-4g' not found`.

This is the qemu upstream cmdline generated by the wrapper, which did
specify max-ram-below-4g, yet `xenpv` was unable to parse it:
```
/usr/lib64/xen/bin/qemu-system-i386 qemu-dm-28 -machine xenpv,max-ram-below-4g=4026531840,allow-unassigned=true,trad_compat=True -xen-domid 28 -vnc 127.0.0.1:1
-device cirrus-vga,vgamem_mb=4,addr=2,romfile=,rombar=1,subvendor_id=0x5853,subsystem_id=0x0001 -vnc-clipboard-socket-fd 4 -xen-domid-restrict -chroot /var/xen/qemu/root-28 -runas 65563.1022
```

`platform:pvfb=true` was an experimental flag that is not documented or
used by XenRT.
(OTOH `other-config:disable_pv_vnc=true` is used by XenRT, which
 disables the console completely for PV guests)

Signed-off-by: Edwin Török <edvin.torok@citrix.com>