CA-304519 wait for QMP socket by connecting to it
It is not obvious for xenopsd when the QMP socket is ready. So far we
only wait for the path to appear but this does not mean it is ready
and trying to connect to it can still fail. Now we are also trying to
establish a connection before returning. If connecting fails, we wait
and try again until the timeout is reached.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>